stddef: fix define for size_t

+++ include/stddef.h
@@ -3,7 +3,7 @@
 // (7.19) Common Definitions
 
 typedef long long int ptrdiff_t;
-typedef long long unsigned int size_t;
+typedef unsigned long int size_t;
 typedef long long int max_align_t;
 
 #define NULL ((void *)0)
